LPS-171473 Simplifying dto conversion
